+++ test/unit/types.ts
@@ -5,6 +5,10 @@ import {
   type Brand,
   type EnvironmentVariables,
   isString,
+  type Nil,
+  type NotNil,
+  type NotNull,
+  type NotUndefined,
   type Phantom,
   type Primitive,
   type TemplateExpression,
@@ -110,6 +114,88 @@ export const typeTests = () => {
         const _: MyType<number> = {
           greet: () => "hello",
         };
+      });
+    });
+
+    suite("Nil", () => {
+      test("should be as expected", () => {
+        expectTypeOf<Nil>().toEqualTypeOf<null | undefined>();
+      });
+    });
+
+    suite("NotNil", () => {
+      test("should work with a non nil type", () => {
+        expectTypeOf<NotNil<number>>().toEqualTypeOf<number>();
+      });
+
+      test("should exclude null from a type", () => {
+        expectTypeOf<NotNil<number | null | string>>().toEqualTypeOf<
+          number | string
+        >();
+      });
+
+      test("should exclude undefined from a type", () => {
+        expectTypeOf<NotNil<number | undefined | string>>().toEqualTypeOf<
+          number | string
+        >();
+      });
+    });
+
+    suite("NotNull", () => {
+      test("should work with a non null type", () => {
+        expectTypeOf<NotNull<number>>().toEqualTypeOf<number>();
+      });
+
+      test("should exclude null from a type", () => {
+        expectTypeOf<NotNull<number | null | string>>().toEqualTypeOf<
+          number | string
+        >();
+      });
+
+      test("should *not* exclude undefined from a type", () => {
+        expectTypeOf<NotNull<number | undefined | string>>().toEqualTypeOf<
+          number | string | undefined
+        >();
+      });
+    });
+
+    suite("NotUndefined", () => {
+      test("should work without unNotUndefined", () => {
+        expectTypeOf<NotUndefined<number>>().toEqualTypeOf<number>();
+      });
+
+      test("should work with number | unNotUndefined", () => {
+        expectTypeOf<
+          NotUndefined<number | undefined>
+        >().toEqualTypeOf<number>();
+      });
+
+      test("should work on an optional field", () => {
+        type Test = {
+          optional?: number;
+        };
+        expectTypeOf<NotUndefined<Test["optional"]>>().toEqualTypeOf<number>();
+      });
+
+      test("should work with an optional field that can be unNotUndefined", () => {
+        type Test = {
+          optional?: number | undefined;
+        };
+        expectTypeOf<NotUndefined<Test["optional"]>>().toEqualTypeOf<number>();
+      });
+
+      test("should work transitively", () => {
+        type First = number | undefined;
+        type Second = First | string | undefined;
+        type Test = Second | undefined;
+
+        expectTypeOf<NotUndefined<Test>>().toEqualTypeOf<number | string>();
+      });
+
+      test("should *not* exlude null from type", () => {
+        expectTypeOf<NotUndefined<number | null>>().toEqualTypeOf<
+          number | null
+        >();
       });
     });