+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.33.0] - 2026-03-11
+
+### Added
+
+- The `Defined` type, which aliases the `NotUndefined` type. This is because double
+negation breaks people's brain. Easier to read this way. Use whichever mfk.
+- The `Nil` type, a union of `null | undefined`.
+- The `NotNil` type, which strips out `null` and `undefined` from any type `T`.
+This is an alias for `NonNullable`. It's shorter, plays well with `es-toolkit`'s
+terminology, and follows `@infra-blocks` naming conventions.
+- The `NotNull` type, which strips out `null` from any type `T`.
+- The `NoUndefined` type which strips out `undefined` from any type `T`.
+
+### Removed
+
+- The `Nullable` type. It conflicts with the new `nil` terminology and can easily
+be replaced by `T | Nil`.
+
 ## [0.32.0] - 2026-03-07
 
 ### Added
@@ -333,6 +351,7 @@ that does the same, but probably better.
   - `isNumber`
   - `isFunction`
 
+[0.33.0]: https://github.com/infra-blocks/ts-types/compare/v0.32.0...v0.33.0
 [0.32.0]: https://github.com/infra-blocks/ts-types/compare/v0.31.0...v0.32.0
 [0.31.0]: https://github.com/infra-blocks/ts-types/compare/v0.30.0...v0.31.0
 [0.30.0]: https://github.com/infra-blocks/ts-types/compare/v0.29.0...v0.30.0